move subscription to userRestController
